+++ src/main/java/br/ufpa/labes/spm/repository/impl/agent/AgentAffinityAgentDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.agent.IAgentAffinityAgentDAO;
 import br.ufpa.labes.spm.domain.AgentAffinityAgent;
 
-public class AgentAffinityAgentDAO extends BaseDAO<AgentAffinityAgent, Integer>
+public class AgentAffinityAgentDAO extends BaseDAO<AgentAffinityAgent, Long>
     implements IAgentAffinityAgentDAO {
 
   protected AgentAffinityAgentDAO(Class<AgentAffinityAgent> businessClass) {
