+++ src/main/java/br/ufpa/labes/spm/repository/impl/agent/AgentDAO.java
@@ -7,7 +7,7 @@ import br.ufpa.labes.spm.repository.interfaces.agent.IAgentDAO;
 import br.ufpa.labes.spm.service.dto.AgentDTO;
 import br.ufpa.labes.spm.domain.Agent;
 
-public class AgentDAO extends BaseDAO<Agent, Integer> implements IAgentDAO {
+public class AgentDAO extends BaseDAO<Agent, Long> implements IAgentDAO {
 
   protected AgentDAO(Class<Agent> businessClass) {
     super(businessClass);
