+++ src/main/java/br/ufpa/labes/spm/repository/impl/agent/OutOfWorkPeriodDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.agent.IOutOfWorkPeriodDAO;
 import br.ufpa.labes.spm.domain.OutOfWorkPeriod;
 
-public class OutOfWorkPeriodDAO extends BaseDAO<OutOfWorkPeriod, Integer>
+public class OutOfWorkPeriodDAO extends BaseDAO<OutOfWorkPeriod, Long>
     implements IOutOfWorkPeriodDAO {
 
   protected OutOfWorkPeriodDAO(Class<OutOfWorkPeriod> businessClass) {
