+++ src/main/java/br/ufpa/labes/spm/repository/impl/artifacts/ArtifactTaskDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.artifacts.IArtifactTaskDAO;
 import br.ufpa.labes.spm.domain.ArtifactTask;
 
-public class ArtifactTaskDAO extends BaseDAO<ArtifactTask, Integer> implements IArtifactTaskDAO {
+public class ArtifactTaskDAO extends BaseDAO<ArtifactTask, Long> implements IArtifactTaskDAO {
 
   protected ArtifactTaskDAO(Class<ArtifactTask> businessClass) {
     super(businessClass);
