+++ src/main/java/br/ufpa/labes/spm/repository/impl/chat/MessageDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.chat.IChatMessageDAO;
 import br.ufpa.labes.spm.domain.ChatMessage;
 
-public class MessageDAO extends BaseDAO<ChatMessage, Integer> implements IChatMessageDAO {
+public class MessageDAO extends BaseDAO<ChatMessage, Long> implements IChatMessageDAO {
 
   protected MessageDAO(Class<ChatMessage> businessClass) {
     super(businessClass);
