+++ src/main/java/br/ufpa/labes/spm/repository/impl/log/CatalogEventDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.log.ICatalogEventDAO;
 import br.ufpa.labes.spm.domain.CatalogEvent;
 
-public class CatalogEventDAO extends BaseDAO<CatalogEvent, Integer> implements ICatalogEventDAO {
+public class CatalogEventDAO extends BaseDAO<CatalogEvent, Long> implements ICatalogEventDAO {
 
   protected CatalogEventDAO(Class<CatalogEvent> businessClass) {
     super(businessClass);
