+++ src/main/java/br/ufpa/labes/spm/repository/impl/log/ProcessEventDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.log.IProcessEventDAO;
 import br.ufpa.labes.spm.domain.ProcessEvent;
 
-public class ProcessEventDAO extends BaseDAO<ProcessEvent, Integer> implements IProcessEventDAO {
+public class ProcessEventDAO extends BaseDAO<ProcessEvent, Long> implements IProcessEventDAO {
 
   protected ProcessEventDAO(Class<ProcessEvent> businessClass) {
     super(businessClass);
