+++ src/main/java/br/ufpa/labes/spm/repository/impl/log/SpmLogDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.log.ISpmLogDAO;
 import br.ufpa.labes.spm.domain.SpmLog;
 
-public class SpmLogDAO extends BaseDAO<SpmLog, Integer> implements ISpmLogDAO {
+public class SpmLogDAO extends BaseDAO<SpmLog, Long> implements ISpmLogDAO {
 
   protected SpmLogDAO(Class<SpmLog> businessClass) {
     super(businessClass);
