+++ src/main/java/br/ufpa/labes/spm/repository/impl/plainActivities/InvolvedArtifactsDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.plainActivities.IInvolvedArtifactsDAO;
 import br.ufpa.labes.spm.domain.InvolvedArtifact;
 
-public class InvolvedArtifactsDAO extends BaseDAO<InvolvedArtifact, Integer>
+public class InvolvedArtifactsDAO extends BaseDAO<InvolvedArtifact, Long>
     implements IInvolvedArtifactsDAO {
 
   protected InvolvedArtifactsDAO(Class<InvolvedArtifact> businessClass) {
