+++ src/main/java/br/ufpa/labes/spm/repository/impl/plainActivities/PrimitiveParamDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.plainActivities.IPrimitiveParamDAO;
 import br.ufpa.labes.spm.domain.PrimitiveParam;
 
-public class PrimitiveParamDAO extends BaseDAO<PrimitiveParam, Integer>
+public class PrimitiveParamDAO extends BaseDAO<PrimitiveParam, Long>
     implements IPrimitiveParamDAO {
 
   protected PrimitiveParamDAO(Class<PrimitiveParam> businessClass) {
