+++ src/main/java/br/ufpa/labes/spm/repository/impl/plainActivities/ReqAgentRequiresAbilityDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.plainActivities.IReqAgentRequiresAbilityDAO;
 import br.ufpa.labes.spm.domain.ReqAgentRequiresAbility;
 
-public class ReqAgentRequiresAbilityDAO extends BaseDAO<ReqAgentRequiresAbility, Integer>
+public class ReqAgentRequiresAbilityDAO extends BaseDAO<ReqAgentRequiresAbility, Long>
     implements IReqAgentRequiresAbilityDAO {
 
   protected ReqAgentRequiresAbilityDAO(Class<ReqAgentRequiresAbility> businessClass) {
