+++ src/main/java/br/ufpa/labes/spm/repository/impl/plannerInfo/ActivityInstantiatedDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.plannerInfo.IActivityInstantiatedDAO;
 import br.ufpa.labes.spm.domain.ActivityInstantiated;
 
-public class ActivityInstantiatedDAO extends BaseDAO<ActivityInstantiated, Integer>
+public class ActivityInstantiatedDAO extends BaseDAO<ActivityInstantiated, Long>
     implements IActivityInstantiatedDAO {
 
   protected ActivityInstantiatedDAO(Class<ActivityInstantiated> businessClass) {
