+++ src/main/java/br/ufpa/labes/spm/repository/impl/plannerInfo/InstantiationPolicyLogDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.plannerInfo.IInstantiationPolicyLogDAO;
 import br.ufpa.labes.spm.domain.InstantiationPolicyLog;
 
-public class InstantiationPolicyLogDAO extends BaseDAO<InstantiationPolicyLog, Integer>
+public class InstantiationPolicyLogDAO extends BaseDAO<InstantiationPolicyLog, Long>
     implements IInstantiationPolicyLogDAO {
 
   protected InstantiationPolicyLogDAO(Class<InstantiationPolicyLog> businessClass) {
