+++ src/main/java/br/ufpa/labes/spm/repository/impl/plannerInfo/ResourcePossibleUseDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.plannerInfo.IResourcePossibleUseDAO;
 import br.ufpa.labes.spm.domain.ResourcePossibleUse;
 
-public class ResourcePossibleUseDAO extends BaseDAO<ResourcePossibleUse, Integer>
+public class ResourcePossibleUseDAO extends BaseDAO<ResourcePossibleUse, Long>
     implements IResourcePossibleUseDAO {
 
   protected ResourcePossibleUseDAO(Class<ResourcePossibleUse> businessClass) {
